Refactor utility methods
